+++ lib/generators/better_translate/install_generator.rb
@@ -4,11 +4,10 @@ module BetterTranslate
   module Generators
     class InstallGenerator < Rails::Generators::Base
       source_root File.expand_path('templates', __dir__)
-
-      desc "Crea il file di configurazione in config/initializers/better_translate.rb"
+      desc "Genera il file config/initializers/better_translate.rb con la configurazione di default"
 
       def copy_initializer
-        template "better_translate_initializer.rb", "config/initializers/better_translate.rb"
+        template "better_translate.rb", "config/initializers/better_translate.rb"
       end
     end
   end